# improve F.rsqrt performance in CPU