`F.rsqrt` optimization for GPU (#4108) made significant performance regression (about 20 times slower) in backward computation when CPU is used.
I separated the code between CPU and GPU to fix this issue.

CPU performance before applying this PR:

* Forward: 1.63ms
* Backward: 70.44ms

CPU performance after applying this PR:

* Forward: 1.75ms
* Backward: 3.71ms

Benchmark code: https://github.com/kmaehashi/chainer-benchmark/blob/master/benchmarks/functions/math/sqrt.py#L20
Data: float32 shape=(1000, 1000)